protecting api routes

+++ frontend/src/components/WorkoutDetails.js
@@ -2,13 +2,18 @@ import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
-
+import { useAuthContext } from '../hooks/useAuthContext'
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext()
-
+ const {user}= useAuthContext();
   const handleClick = async () => {
+
+    if(!user){
+       return ;
+    }
     const response = await fetch('/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers:{'Authorization': `Bearer ${user.token}`}
     })
     const json = await response.json()
